requirements.txt
# PubSub and GCS API
google-cloud-storage==1.8.0
google-cloud-pubsub==0.34.0
google-cloud-bigquery==1.0.0
google-cloud-bigquery-datatransfer==0.1.1
google-api-python-client==1.6.6

google-auth-httplib2==0.0.3
path.py==11.0.1
ruamel.yaml==0.15.37
protobuf3-to-dict==0.1.5